✨ Move paywall logic to separate view model
