+++ ios/PresentationLayer/Integrations/Sources/Integrations/Overview/IntegrationsOverviewViewModel.swift
@@ -22,15 +22,20 @@ final class IntegrationsOverviewViewModel: BaseViewModel, ViewModel, ObservableO
     
     @Injected(\.getIntegrationsUseCase) private var getIntegrationsUseCase
     @Injected(\.getHasFullAccessUseCase) private var getHasFullAccessUseCase
-    @Injected(\.getPurchasePackagesUseCase) private var getPurchasePackagesUseCase
-    @Injected(\.purchasePackageUseCase) private var purchasePackageUseCase
-    @Injected(\.getIsPackageEligibleForIntroductoryDiscountUseCase) private var getIsPackageEligibleForIntroductoryDiscountUseCase
     
     private weak var flowController: FlowController?
     
+    // MARK: - Stored properties
+    
+    let paywallViewModel: PaywallViewModel
+    
     // MARK: - Init
     
-    init(flowController: FlowController? = nil) {
+    init(
+        paywallViewModel: PaywallViewModel,
+        flowController: FlowController? = nil
+    ) {
+        self.paywallViewModel = paywallViewModel
         self.flowController = flowController
         super.init()
     }
@@ -54,8 +59,6 @@ final class IntegrationsOverviewViewModel: BaseViewModel, ViewModel, ObservableO
         var integrations: ViewData<[Integration]> = .loading(mock: .stub)
         var isShowingTypes = false
         var showPaywall: ViewData<Bool> = .loading(mock: false)
-        var purchasePackages: ViewData<[PaywallPackageViewObject]> = .loading(mock: .stub)
-        var purchaseLoading = false
     }
     
     // MARK: - Intent
@@ -66,7 +69,6 @@ final class IntegrationsOverviewViewModel: BaseViewModel, ViewModel, ObservableO
         case selectNewIntegrationType(IntegrationType)
         case showIntegrationDetail(id: String)
         case changeShowingTypes(to: Bool)
-        case purchasePackage(packageId: String)
     }
     
     func onIntent(_ intent: Intent) {
@@ -77,7 +79,6 @@ final class IntegrationsOverviewViewModel: BaseViewModel, ViewModel, ObservableO
             case let .showIntegrationDetail(id): showIntegrationDetail(integrationId: id)
             case let .changeShowingTypes(isShowing): state.isShowingTypes = isShowing
             case let .selectNewIntegrationType(type): selectNewIntegrationType(type)
-            case let .purchasePackage(packageId): await purchasePackage(packageId: packageId)
             }
         })
     }
@@ -96,8 +97,6 @@ final class IntegrationsOverviewViewModel: BaseViewModel, ViewModel, ObservableO
             
             if hasFullAccess.boolValue {
                 await fetchIntegrations(showLoading: showLoading)
-            } else {
-                await fetchPaywall(showLoading: showLoading)
             }
         } onError: { error in
             state.integrations = .error(error)
@@ -139,48 +138,6 @@ final class IntegrationsOverviewViewModel: BaseViewModel, ViewModel, ObservableO
             delegate: self
         )))
     }
-    
-    private func fetchPaywall(showLoading: Bool) async {
-        if showLoading {
-            state.purchasePackages = .loading(mock: .stub)
-        }
-        
-        await execute {
-            let packages: [PurchasePackage] = try await getPurchasePackagesUseCase.execute()
-            let viewObjects = try await packages.asyncMap { package in
-                let params = GetIsPackageEligibleForIntroductoryDiscountUseCaseParams(
-                    packageId: package.id
-                )
-                
-                return PaywallPackageViewObject(
-                    package: package,
-                    isEligibleForIntroductoryDiscount: try await getIsPackageEligibleForIntroductoryDiscountUseCase.execute(
-                        params: params
-                    )
-                )
-            }
-            
-            if packages.isEmpty {
-                state.purchasePackages = .empty(.noData)
-            } else {
-                state.purchasePackages = .data(viewObjects)
-            }
-        } onError: { error in
-            state.purchasePackages = .error(error)
-        }
-    }
-    
-    private func purchasePackage(packageId: String) async {
-        state.purchaseLoading = true
-        defer { state.purchaseLoading = false }
-        
-        await execute {
-            let params = PurchasePackageUseCaseParams(packageId: packageId)
-            try await purchasePackageUseCase.execute(params: params)
-        } onError: { error in
-            state.showPaywall = .error(error)
-        }
-    }
 }
 
 // MARK: IntegrationDetailViewModelDelegate
@@ -197,5 +154,15 @@ extension IntegrationsOverviewViewModel: IntegrationDetailViewModelDelegate {
                 message: "\(L10n.integration_view_removed_snack_title_part_one) \(integrationName) \(L10n.integration_view_removed_snack_title_part_two)"
             ))
         }
+    }
+}
+
+// MARK: PaywallViewModelDelegate
+
+extension IntegrationsOverviewViewModel: PaywallViewModelDelegate {
+    func didPurchasePackage(_ package: PurchasePackage) {
+        executeTask(Task{
+            await fetchData(showLoading: true)
+        })
     }
 }
