+++ ios/PresentationLayer/UIToolkit/Package.swift
@@ -19,6 +19,7 @@ let package = Package(
         // .package(url: /* package url */, from: "1.0.0"),
         .package(name: "Utilities", path: "../../DomainLayer/Utilities"),
         .package(name: "SharedDomain", path: "../../DomainLayer/SharedDomain"),
+        .package(name: "DependencyInjection", path: "../../Application/DependencyInjection"),
         .package(url: "https://github.com/SwiftGen/SwiftGenPlugin", .upToNextMajor(from: "6.6.0")),
         .package(url: "https://github.com/SFSafeSymbols/SFSafeSymbols.git", exact: "5.2.0")
     ],
@@ -30,6 +31,8 @@ let package = Package(
             dependencies: [
                 .product(name: "Utilities", package: "Utilities"),
                 .product(name: "SharedDomain", package: "SharedDomain"),
+                .product(name: "DependencyInjection", package: "DependencyInjection"),
+                .product(name: "DependencyInjectionMocks", package: "DependencyInjection"),
                 .product(name: "SharedDomainMocks", package: "SharedDomain"),
                 .product(name: "SFSafeSymbols", package: "SFSafeSymbols")
             ],
