+++ ios/PresentationLayer/UIToolkit/Sources/UIToolkit/BaseViews/SwiftUI/Paywall/ViewObjects/PaywallViewOrigin.swift
@@ -1,0 +1,9 @@
+//
+//  Created by Tomáš Batěk on 20.11.2024
+//  Copyright © 2024 Matee. All rights reserved.
+//
+
+public enum PaywallViewOrigin {
+    case generic
+    case integrations
+}
